# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Chavez API
  version: 0.1.0
  description: |
    API del harness Chavez: auth (Better Auth + magic link + device), vault de providers,
    workspaces/sessions/chats por HTTP, y sincronización vía WebSocket `/ws`.

    **Auth:** Bearer (`Authorization: Bearer <session-token>`) o cookie de sesión Better Auth.
    Sign-in y sign-up usan el mismo email (magic link).

    **WebSocket:** conectar a `ws://<host>/ws?token=<token>` y enviar mensajes JSON
    `{ type, id, ... }` → respuesta `{ type, id, ok, data?, error? }`.

    Las rutas bajo `/api/auth/*` las atiende Better Auth (catch-all). Los shapes documentados
    reflejan el uso en este proyecto; detalles pueden variar con la versión del plugin.
  contact:
    name: Chavez
servers:
  - url: http://localhost:25001
    description: Stack local (API fija :25001)
...
paths:
...
  /providers/:
    get:
      tags: [Providers]
      summary: Listar providers, prefs y catálogos
      description: |
        models son nativos por provider; raw es el JSON recolectado.
        Query `refresh=1` fuerza rediscovery de Cursor. 401 sin sesión.
      operationId: listProviders
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: refresh
          in: query
          required: false
          schema:
            type: string
            enum: ["1"]
          description: Fuerza refresh del catálogo Cursor aunque el cache esté fresco.
      responses:
        "200":
          description: Estado de vault + catálogos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvidersListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /providers/preferences:
    put:
      tags: [Providers]
      summary: Actualizar preferencias activas
      operationId: putProviderPreferences
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderPreferencesBody"
      responses:
        "200":
          description: Preferencias actualizadas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderPreferences"
        "400":
          description: |
            Request inválido. Si `activeExecutionMode` no es `plan`, `auto` o `ask`,
            el body `error` es `executionMode must be plan, auto, or ask` y la fila no cambia.
            Si `activeModel` no pertenece a `activeProvider`, el body `error` nombra ambos
            (`modelId "…" does not belong to provider …`). Params Cursor inválidos usan
            `param "id=value" is not valid for model …`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: executionMode must be plan, auto, or ask
        "401":
          $ref: "#/components/responses/Unauthorized"
  /providers/active:
    put:
      tags: [Providers]
      summary: Fijar provider activo
      operationId: putActiveProvider
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider]
              properties:
                provider:
                  oneOf:
                    - $ref: "#/components/schemas/ProviderId"
                    - type: "null"
      responses:
        "200":
          description: Provider activo (modelo/params recortados al catálogo)
          content:
            application/json:
              schema:
                type: object
                properties:
                  activeProvider:
                    oneOf:
                      - $ref: "#/components/schemas/ProviderId"
                      - type: "null"
                  activeModel:
                    type: string
                    nullable: true
                  activeEffort:
                    type: string
                    nullable: true
                  activeParams:
                    type: array
                    nullable: true
                    items:
                      $ref: "#/components/schemas/CursorParamSelection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /providers/{provider}/credentials:
    parameters:
      - $ref: "#/components/parameters/ProviderParam"
    put:
      tags: [Providers]
      summary: Guardar / actualizar credenciales (vault)
      operationId: putProviderCredentials
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [authKind, secret]
              properties:
                authKind:
                  $ref: "#/components/schemas/AuthKind"
                secret:
                  type: string
                  minLength: 1
      responses:
        "200":
          description: Credenciales guardadas
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  provider:
                    $ref: "#/components/schemas/ProviderId"
                  authKind:
                    $ref: "#/components/schemas/AuthKind"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      tags: [Providers]
      summary: Leer credenciales desencriptadas
      operationId: getProviderCredentials
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Secret en claro (uso CLI / hub autenticado)
          content:
            application/json:
              schema:
                type: object
                properties:
                  provider:
                    $ref: "#/components/schemas/ProviderId"
                  authKind:
                    $ref: "#/components/schemas/AuthKind"
                  secret:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Providers]
      summary: Eliminar credenciales
      operationId: deleteProviderCredentials
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Eliminado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: session-token
      description: Token de sesión Better Auth (CLI device token o session token)
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: |
        Cookie de sesión Better Auth (nombre exacto puede incluir prefijo/host).
        El hub web usa `credentials: include`.
  parameters:
    ProviderParam:
      name: provider
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ProviderId"
...
  responses:
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            error: Unauthorized
    BadRequest:
      description: Request inválido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          example: true
    ErrorBody:
      type: object
      properties:
        error:
          type: string
...
    ProviderId:
      type: string
      enum: [claude, cursor]
    AuthKind:
      type: string
      enum: [oauth_token, api_key]
...
    CursorParamSelection:
      type: object
      required: [id, value]
      properties:
        id:
          type: string
        value:
          type: string
...
    ProviderCatalog:
      type: object
      description: |
        models son nativos por provider (Claude → ModelInfo con effortLevels;
        Cursor → CursorModelInfo con parameters/variants). raw es el JSON recolectado.
      properties:
        id:
          $ref: "#/components/schemas/ProviderId"
        label:
          type: string
        models:
          type: array
          items: {}
        raw: {}
        catalogError:
          type: string
          nullable: true
        runnable:
          type: boolean
    ProviderStatus:
      type: object
      properties:
        linked:
          type: boolean
        authKind:
          $ref: "#/components/schemas/AuthKind"
        updatedAt:
          type: string
          format: date-time
        label:
          type: string
        runnable:
          type: boolean
        catalogError:
          type: string
          nullable: true
        models:
          type: array
          items: {}
    ProvidersListResponse:
      type: object
      description: models son nativos por provider; raw es el JSON recolectado.
      properties:
        activeProvider:
          oneOf:
            - $ref: "#/components/schemas/ProviderId"
            - type: "null"
        activeModel:
          type: string
          nullable: true
        activeEffort:
          type: string
          nullable: true
        activeParams:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/CursorParamSelection"
        activeExecutionMode:
          type: string
          nullable: true
          enum: [plan, auto, ask]
        catalogs:
          type: array
          items:
            $ref: "#/components/schemas/ProviderCatalog"
        providers:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ProviderStatus"
    ProviderPreferencesBody:
      type: object
      properties:
        activeProvider:
          oneOf:
            - $ref: "#/components/schemas/ProviderId"
            - type: "null"
        activeModel:
          type: string
          nullable: true
        activeEffort:
          type: string
          nullable: true
        activeParams:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/CursorParamSelection"
        activeExecutionMode:
          type: string
          nullable: true
          enum: [plan, auto, ask]
    ProviderPreferences:
      type: object
      properties:
        activeProvider:
          type: string
          nullable: true
        activeModel:
          type: string
          nullable: true
        activeEffort:
          type: string
          nullable: true
        activeParams:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/CursorParamSelection"
        activeExecutionMode:
          type: string
          nullable: true
          enum: [plan, auto, ask]
...